Added MachineID tracking
Updated the ban so that it is more robust in keeping people banned. the client sends hardware id that can be added to the database. adding that to the players account insures that once banned, alt account and new account still wont be able to login. not bulletproof but will help. maybe ahs potential other uses